Basic network mapping
A basic network mapping for a street by using  linked list data structure. The Internet comes with cables until
wireless modem in houses,
and there is a bandwidth value which is related to how much money you is paid for the Internet connection for each month. A basic Internet connection mapping in a
street is created in that project.